+++ plastron-models/src/plastron/models/ldp.py
@@ -1,0 +1,7 @@
+from plastron.namespaces import ldp
+from plastron.rdfmapping.descriptors import ObjectProperty
+from plastron.rdfmapping.resources import RDFResource
+
+
+class LDPContainer(RDFResource):
+    contains = ObjectProperty(ldp.contains, repeatable=True)
